refactor(testing_server): Move testing_server fixture into seperate file for sharing among tests.
